feat(deps): drop OpenSSL dependency in favor of Kaesar
Normally OpenSSL and other variants of this kind are a sturdy choice. My goal
is not to rely on random binaries around and instead use built-in Emacs
functionality, or in the worst case, ELisp libraries.

For the moment I don't have a need to write AES CBC from scratch and Kaesar
fits the needs as a well-replacable func call for `openssl enc -aes-256-cbc`
just fine.